pyproject.toml
[project]
name = "clage"
version = "0.3.0"
description = "Evolutionary simulation with neural networks"
authors = [
    {name = "Your Name", email = "your@email.com"}
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "pygame-ce>=2.0",
    "numpy>=1.0",
    "matplotlib>=3.0",
    "tqdm>=4.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "black>=23.0",
    "flake8>=6.0",
    "pylint>=3.0"
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
python_files = "test_*.py"
addopts = "-v --cov=./ --cov-report=html"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py310']

py-modules = ["evolve_sim", "grid", "main", "genome"]